requirements.txt
# BotBuddy - Multi-Model Edition (100% Free)

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client for Ollama
httpx==0.26.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Optional - Anthropic fallback (uncomment if you want cloud backup)
# anthropic==0.18.1

# Optional - Voice (uncomment if needed)
# elevenlabs==1.0.0
# assemblyai==0.23.0

# Optional - SMS (uncomment if needed)
# twilio==8.13.0
